% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lxb_data.R
\name{analyseBeadDistributions}
\alias{analyseBeadDistributions}
\title{Plot the distribution of beads in each well}
\usage{
analyseBeadDistributions(lxb_dataset, region, analyte = "", tpw = list())
}
\arguments{
\item{region}{Number of the beads to analyse}

\item{analyte}{Name of the analytes corresponding to the bead number of region}

\item{tpw}{List of the treatments applied to each well (names(tpw) is the name of the well)}
}
\description{
Plot the distribution of beads in each well
}

