% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lxb_data.R
\name{computeVariation}
\alias{computeVariation}
\title{Compute the coefficient of variation for one readout}
\usage{
computeVariation(wells_per_treatment, bead_carac, wilcox_sample, median_cv,
  blanks, controls, default_cv = 0.3, min_cv = 0.1)
}
\arguments{
\item{wells_per_treatment}{List of vectors, each entry of the list is named after a treatment and contains a list of wells coordinates}

\item{bead_carac}{Values for all the wells for one bead}

\item{wilcox_sample}{List of the distribution of values of the bead in each well}
}
\description{
Compute the coefficient of variation for one readout taking into account the variation of the for each well and the variation between samples
}

