% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lxb_data.R
\name{extractExperimentalSetup}
\alias{extractExperimentalSetup}
\title{Get the association between wells and treatments}
\usage{
extractExperimentalSetup(plate)
}
\arguments{
\item{plate}{An array describing the organisation of the plate. The empty string "" indicates that no treatment were used and that informations should not be extracted from the well.}
}
\description{
Get the association between wells and treatments
}
