% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lxb_data.R
\name{processReadout}
\alias{processReadout}
\title{Extract data for a bead number in a well from an lxb dataset}
\usage{
processReadout(dataset, bead, well, PLOT = FALSE)
}
\arguments{
\item{dataset}{An full lxb dataset}

\item{bead}{The bead number}

\item{well}{The well identifier}
}
\description{
Select valid data for the specified bead of the specified well and perform a bootstrap to extract a robust mean an the variation
}

