% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lxb_data.R
\name{readBeads}
\alias{readBeads}
\title{Process the bead dataset}
\usage{
readBeads(lxb_dataset, region, analyte, controls, blanks,
  wells_per_treatment_with_blank, plot_dir = "", externals = c(""),
  PLOT = F)
}
\arguments{
\item{lxb_dataset}{The lxb dataset}

\item{region}{Beads IDs used}

\item{analyte}{Analyte corresponding to the analysis.}

\item{controls}{Wells containing control conditions.}

\item{blanks}{Wells containing no cell.}

\item{wells_per_treatment_with_blank}{List of wells where each treatment is applied. It is a list where each entry is named after a treatment and contains a vector of well names.}

\item{externals}{Wells to exclude from the processing}

\item{PLOT}{Whether data should be plotted}
}
\description{
Process the bead dataset
}

