% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lxb_data.R
\name{sortBeads}
\alias{sortBeads}
\title{Extract values from the lxb dataset}
\usage{
sortBeads(lxb_dataset, wells_per_treatment_with_blank, region, analyte)
}
\arguments{
\item{wells_per_treatment_with_blank}{List of wells where each treatment is applied. It is a list where each entry is named after a treatment and contains a vector of well names.}

\item{region}{Beads IDs used}

\item{analyte}{Analyte corresponding to the analysis.}

\item{dataset}{An full lxb dataset}
}
\description{
Extract values from the lxb dataset in a nicely formated list
}
