% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lxb_data.R
\name{writeMIDAS}
\alias{writeMIDAS}
\title{Write lxb data in MIDAS format}
\usage{
writeMIDAS(beads, dname, inhibitors, stimulators, midas_dir = "midas_data")
}
\arguments{
\item{beads}{Beads data}
}
\description{
Write data extracted and processed using readBeads in MIDAS format
}
