% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lxb_data.R
\name{writeMIDASfile}
\alias{writeMIDASfile}
\title{Write the data in a file in MIDAS format}
\usage{
writeMIDASfile(datas, variations, dname, blanks, controls, wells_per_treatment,
  inhibitors, stimulators, midas_dir = "midas_data")
}
\arguments{
\item{datas}{The values measured. Should a matrix with the wells as rownames and the analyte as column names}

\item{variations}{The coefficient of variation from the replicates}

\item{dname}{The name of the data}
}
\description{
Write the data in a file in MIDAS format
}
