% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lxb_data.R
\name{extractExperimentalSetup}
\alias{extractExperimentalSetup}
\title{Get the association between wells and treatments}
\usage{
extractExperimentalSetup(plate)
}
\arguments{
\item{plate}{An array describing the organisation of the plate, must be defined in the R session}
}
\description{
Get the association between wells and treatments
}

